requirements.txt
prefect>=2.0.0
pandas>=1.5.0
yfinance>=0.2.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
streamlit>=1.20.0
vectorbt>=0.24.0
python-dotenv>=1.0.0
httpx>=0.24.0
matplotlib>=3.7.0
